Create save_game and load_game methods
This commit creates the save_game and load_game methods.

Method save_game takes the word, secret and wrong_guess_counter
instance variables and turn then into a JSON file.

Method load_game takes the JSON file and load them back at the matching
variables.

I chose JSON because it's a format I work with, and I wanted to see how
it would behave using Ruby. There is a blatantly flaw, which is that
the player can easily open the JSON file to see what is the word.